Refactor :  move onBoarding logic to new screen
